Refactor orchestration and validation components
- Introduced parsing helpers for converting mappings into typed domain documents in `parsing.py`.
- Enhanced curriculum and topic specification handling in `planning.py` with type-safe functions.
- Added scope policy management in `scope_policy.py` to define synthesis behavior based on profiles.
- Implemented a new artifact management system in `scope_artifacts.py` for versioned scope artifacts.
- Reorganized validation logic into a rule-based system in `rules.py` for better modularity and maintainability.
- Updated orchestration pipeline in `pipeline.py` to integrate new scope and validation features.
- Added comprehensive tests for orchestration commands and planning precision to ensure functionality.

+++ learning_compiler/domain/models.py
@@ -167,7 +167,7 @@ class CurriculumNode:
     core_ideas: tuple[str, ...]
     pitfalls: tuple[str, ...]
     mastery_check: MasteryCheck
-    estimate_minutes: int
+    estimate_minutes: float
     resources: tuple[Resource, ...]
     estimate_confidence: float | None = None
